Fix fatal class redeclaration by renaming endpoint class to Jetpack_Launch_Site
Rename WPCOM_REST_API_V2_Endpoint_Launch_Site to
WPCOM_REST_API_V2_Endpoint_Jetpack_Launch_Site to avoid a fatal conflict
with the identically named class in WPCOM core's rest-api-plugins.
